[Tests] deactivate a broken serialization check
We had some nice serialized data, and intended to ensure they will be
future proof by unit-testing the deserialization.

But pinocchio v2.6.0 break this.

By deactivating this test, we acknowledge that the serialization feature
of this project is mostly intended to exchange data between simultaneous
processes, and not to store data or results in the long term.

This commit just fix the CI, but we should also remove parts that are
now useless, and properly document the goal of this project.